requirements.txt
fastapi
uvicorn
openai
beautifulsoup4
requests
numpy
pydantic
python-dotenv
